Update generated file to remove asciidoc from name
Files would get generated as original filename with just the
html extension added to them.

The index.hml when opened would try to point to the files by
name, but without the intermediate asciidoc extension.

Updated to have the build.sh script to drop the asciidoc extension
from the generated file and just have the extension be just html.

e.g. 01_rebar.asciidoc previously got generated as 01_rebar.asciidoc,
but is now generated as 01_rebar.html